+++ apps/data/models/__init__.py
@@ -1,6 +1,7 @@
 from .checklist import Checklist
 from .country import Country
 from .county import County
+from .filter import Filter
 from .location import Location
 from .observation import Observation
 from .observer import Observer
@@ -11,6 +12,7 @@ __all__ = (
     "Checklist",
     "Country",
     "County",
+    "Filter",
     "Location",
     "Observation",
     "Observer",