+++ CHANGELOG.md
@@ -8,6 +8,7 @@ a concise history of changes made.
 ## Latest
 - Added an About page, describing the site.
 - Added a field to the Species page, so you can search for species, subspecies, etc.
+- Added a Filter model for mapping Guineafowls, etc. to their domestic form.
 - Fixed error where the species total was also counting subspecies, forms, etc.
 
 ## 2025.06.10